Fix aliasing in identify method.
The identify mapper method spits out an alias field that is formatted for the
KISSMetrics API. e.g.

```
alias = {
      _k: this.settings.apiKey,
      _p: identify.userId(),
      _n: identify.sessionId()
}
```

The identify method incorrectly calls alias with this KISSMetrics
formatted object, instead of a Segment formatted alias call.

This causes this line `_p: alias.from(),` in the alias method mapper to
fail, because it is not being passed an alias facade object.

This patch fixes the identify method so that it directly makes a request
with the output of the identify mapper rather than running it through
the alias mapper.

This bug slipped through because the fixture identify-alias was only
being tested for mapping, but not being run through the integration.
This patch adds a test case that verifies an identify call makes 2
successful requests to the KISSMetrics API.